Centralized learning/requirements.txt
certifi==2020.6.20
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
urllib3==1.25.9
flask==1.1.2
gunicorn==20.0.4
gevent==20.6.2
paho-mqtt==1.5.0
Celery==4.4.7

numpy~=1.21.6
scipy~=1.6.2
pandas~=1.3.5
keras~=2.3.1
scikit-learn~=0.19.2
tqdm~=4.64.1
seaborn~=0.12.2
coremltools~=4.1
zlib~=1.2.13
pip~=22.3.1
attrs~=22.1.0
wheel~=0.38.4
# openssl~=1.1.1u
cryptography~=39.0.1
pillow~=9.4.0
jinja2~=3.1.2
setuptools~=65.6.3
packaging~=23.0
pyparsing~=3.0.9
zipp~=3.13.0
typing_extensions~=4.4.0
readline~=7.0
torchvision~=0.13.1
wandb~=0.13.10
mpi4py~=3.1.4
networkx~=2.2
yaml~=0.2.5
pyyaml~=6.0
wfdb~=4.1.0
psutil~=5.9.4
h5py~=3.7.0
setproctitle~=1.3.2
pyautogui~=0.9.54